Improve making symlinks clearly

+++ symlink_dotfiles.sh
@@ -3,9 +3,12 @@ for i in {.awsrc,.bash_profile,.bashrc,.gemrc,.gitconfig,.gitexclude,.rspec,.vim
 do
 	if [ ! -f $PWD/$i -a -f $PWD/$i.sample ]; then
 		cp $PWD/$i.sample $PWD/$i
-		echo "$PWD/$i.sample $PWD/$i"
+		echo "copied $PWD/$i.sample $PWD/$i"
 	fi
-	mv ~/$i ~/$i.original
+	if [ -f ~/$i ]; then
+		mv ~/$i ~/$i.original
+		echo "moved ~/$i ~/$i.original"
+	fi
 	ln -s $PWD/$i ~/$i
-	echo "ln -s $PWD/$i ~/$i"
+	echo "linked $PWD/$i ~/$i"
 done
